home page added to navbar
