+++ src/components/HomePage/index.js
@@ -1,6 +1,5 @@
-import React from "react";
+import React, { useState } from "react";
 import Card from "@material-ui/core/Card";
-import { makeStyles } from "@material-ui/core/styles";
 import CardComponent from "../util/CodelabCard/index";
 import Typography from "@material-ui/core/Typography";
 import BottomNavigation from "@material-ui/core/BottomNavigation";
@@ -11,91 +10,14 @@ import ListItem from "@material-ui/core/ListItem";
 import ListItemText from "@material-ui/core/ListItemText";
 import List from "@material-ui/core/List";
 import { userList } from "./userList";
+import useStyles from "./styles";
 
 function HomePage() {
-  const useStyles = makeStyles((theme) => ({
-    wrapper: {
-      display: "flex",
-      alignItems: "top",
-      justifyContent: "center",
-      height: "100%",
-
-      background: "#f2f2f2",
-    },
-    mainBody: {
-      display: "flex",
-      alignContent: "center",
-      justifyContent: "center",
-      //   flex: "1",
-      height: "100%",
-      margin: "1rem 0 2rem 0",
-      flexDirection: "column",
-    },
-    sideBody: {
-      display: "flex",
-      alignContent: "flex-start",
-      justifyContent: "flex-start",
-      flex: "1",
-      marginTop: "5rem",
-      maxWidth: "20rem",
-      margin: "0 1rem 2rem 1rem",
-      height: "100%",
-      flexDirection: "column",
-
-      [theme.breakpoints.down(750)]: {
-        display: "none",
-      },
-    },
-    cardBody: {
-      display: "flex",
-      justifyContent: "space-between",
-      width: "100%",
-      alignItems: "center",
-      direction: "column",
-    },
-    pp: {
-      display: "flex",
-      justifyContent: "center",
-      padding: "5rem",
-      background: "red",
-    },
-    sort: {
-      width: "100%",
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "space-between",
-      flexDirection: "row",
-      margin: "0rem 0 0rem 0",
-    },
-    sortedList: {
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "space-around",
-      flexDirection: "row",
-      width: "auto",
-      [theme.breakpoints.down(750)]: {
-        display: "none",
-      },
-    },
-    navigation: {
-      "&:selcted": {
-        border: "2px solid black",
-      },
-    },
-    sideCard: {
-      minWidth: "20vw",
-      display: "flex",
-      alignItems: "center",
-      justifyContent: "center",
-      maxHeight: "30rem",
-      overflow: "auto",
-      margin: "0 0 2rem 0",
-      background: "white",
-      boxShadow: ".5px 2px 5px gray",
-    },
-  }));
   const classes = useStyles();
-
+  const [value, setValue] = useState(2);
+  const handleChange = (event, newValue) => {
+    setValue(newValue);
+  };
   return (
     <Card className={classes.wrapper}>
       <div className={classes.sideBody}>
@@ -104,7 +26,11 @@ function HomePage() {
           className={classes.sideCard}
           alignContent="center"
           direction="column"
-          style={{ padding: "1rem" }}
+          style={{
+            width: "100%",
+            overflow: "auto",
+            maxHeight: "25rem",
+          }}
         >
           <Grid item>
             <Typography
@@ -118,7 +44,6 @@ function HomePage() {
           <Grid container alignItems="left">
             <List
               component="nav"
-              className={classes.root}
               aria-label="mailbox folders"
               style={{ width: "100%" }}
             >
@@ -144,7 +69,11 @@ function HomePage() {
           className={classes.sideCard}
           alignContent="center"
           direction="column"
-          style={{ padding: "1rem" }}
+          style={{
+            width: "100%",
+            overflow: "auto",
+            maxHeight: "25rem",
+          }}
         >
           <Grid item>
             <Typography
@@ -158,7 +87,6 @@ function HomePage() {
           <Grid container alignItems="left">
             <List
               component="nav"
-              className={classes.root}
               aria-label="mailbox folders"
               style={{ width: "100%" }}
             >
@@ -173,6 +101,7 @@ function HomePage() {
                 <ListItemText primary="Lorem Pervious Text" />
               </ListItem>
               <Divider light />
+
               <ListItem button>
                 <ListItemText primary="Lorem Pervious Text Previous Event" />
               </ListItem>
@@ -182,18 +111,20 @@ function HomePage() {
       </div>
       <div className={classes.mainBody}>
         <Grid container className={classes.sort}>
-          <Typography variant="h6" style={{ padding: ".5rem 2rem" }}>
-            Tutorial
+          <Typography
+            variant="h6"
+            style={{ padding: ".5rem 1rem", fontWeight: "bold" }}
+          >
+            Posts
           </Typography>
           <BottomNavigation
             showLabels
             style={{ background: "#f2f2f2" }}
             className={classes.sortedList}
+            value={value}
+            onChange={handleChange}
           >
-            <BottomNavigationAction
-              label="week"
-              className={classes.navigation}
-            ></BottomNavigationAction>
+            <BottomNavigationAction label="week" />
             <BottomNavigationAction
               label="Month"
               style={{ fontSize: "2rem" }}
@@ -232,9 +163,12 @@ function HomePage() {
           <Grid container alignItems="left">
             <List
               component="nav"
-              className={classes.root}
               aria-label="mailbox folders"
-              style={{ width: "100%" }}
+              style={{
+                width: "100%",
+                overflow: "auto",
+                maxHeight: "25rem",
+              }}
             >
               <ListItem button>
                 <ListItemText primary="lorem lorem text lorem text text lorem text lorem text" />
@@ -260,29 +194,31 @@ function HomePage() {
           direction="column"
           style={{ padding: "1rem" }}
         >
-          <Grid item>
-            <Typography
-              variant="h6"
-              gutterBottom
-              style={{ marginBottom: "1rem" }}
-            >
-              Discussion
-            </Typography>
-          </Grid>
+          <Typography
+            variant="h6"
+            gutterBottom
+            style={{ marginBottom: "1rem" }}
+          >
+            Discussion
+          </Typography>
           <Grid container alignItems="left">
             <List
               component="nav"
-              className={classes.root}
               aria-label="mailbox folders"
-              style={{ width: "100%" }}
+              style={{
+                width: "100%",
+                overflow: "auto",
+                maxHeight: "25rem",
+              }}
             >
               <ListItem button>
                 <ListItemText primary="lorem lorem text lorem text text lorem text lorem text" />
               </ListItem>
-              <Divider />
-              <ListItem button divider>
+              <Divider light />
+              <ListItem button>
                 <ListItemText primary="lorem lorem text lorem text text lorem text lorem text" />
               </ListItem>
+              <Divider light />
               <ListItem button>
                 <ListItemText primary="lorem lorem text lorem text text lorem text lorem text" />
               </ListItem>
